Refactored code to organize all resumio sections with repeatable subSections.  Reorganized folder structure to make more sense.  nested sections are now segments.
