--- frontend/src/components/control/ResumioApp/InputApp/SectionComponents/LanguagesI.jsx
@@ -1,33 +1,0 @@
-import React from 'react'
-import { useFormContext } from 'react-hook-form'
-import {ResumioLabel, 
-        ResumioInput,
-        SectionFieldset} from '../../../../style/ResumioApp/InputApp/SectionStyles'
-
-function LanguagesI({sectionData, registrationStr}) {
-  const {register} = useFormContext()
-  return (
-    <>
-      {sectionData.map((language, index)=> {
-        return(
-          <SectionFieldset key={language.language}>
-            <ResumioLabel htmlFor="language">Language</ResumioLabel>
-            <ResumioInput 
-              {...register(registrationStr + `${index}.language`)} 
-              name="language" 
-              defaultValue={language["language"]} 
-            />
-            <ResumioLabel htmlFor="fluency">Fluency</ResumioLabel>
-            <ResumioInput 
-              {...register(registrationStr + `${index}.fluency`)} 
-              name="fluency" 
-              defaultValue={language["fluency"]} 
-            />
-          </SectionFieldset>
-        )
-      })}
-    </>
-  )
-}
-
-export default LanguagesI
