--- frontend/src/components/control/ResumioApp/InputApp/SectionComponents/PublicationsI.jsx
@@ -1,56 +1,0 @@
-import React from 'react'
-import { useFormContext } from 'react-hook-form'
-import {ResumioLabel, 
-        ResumioInput,
-        ResumioTextArea, 
-        SectionLegend,
-        SectionFieldset} from '../../../../style/ResumioApp/InputApp/SectionStyles'
-
-function PublicationsI({sectionData,registrationStr}) {
-  const {register} = useFormContext()
-  return (
-    <>
-      {sectionData.map((publication, index) => {
-        return(
-          <SectionFieldset key={`${publication.name}_${publication.releaseDate}`}>
-            <SectionLegend>{publication.name}</SectionLegend>
-            <ResumioLabel htmlFor="name">Publication Title</ResumioLabel>
-            <ResumioInput 
-              {...register(registrationStr +`${index}.name`)} 
-              name="name" 
-              defaultValue={publication["name"]}
-            />
-            <ResumioLabel htmlFor="name">Publisher</ResumioLabel>
-            <ResumioInput 
-              {...register(registrationStr +`${index}.publisher`)} 
-              name="publisher" 
-              defaultValue={publication["publisher"]}
-            />
-            <ResumioLabel htmlFor="releaseDate">Publication Date</ResumioLabel>
-            <ResumioInput 
-              {...register(registrationStr +`${index}.releaseDate`)}
-              type='date' 
-              name="releaseDate" 
-              defaultValue={publication["releaseDate"]}
-            />
-            <ResumioLabel htmlFor="summary">Synopsis</ResumioLabel>
-            <ResumioTextArea 
-            rows='4'
-            {...register(registrationStr +`${index}.summary`)} 
-            name="summary" 
-            defaultValue={publication["summary"]} 
-            />
-            <ResumioLabel htmlFor="url">URL</ResumioLabel>
-            <ResumioInput 
-              {...register(registrationStr +`${index}.url`)}
-              name="url" 
-              defaultValue={publication["url"]}
-            />
-          </SectionFieldset>
-        )
-      })}
-    </>
-  )
-}
-
-export default PublicationsI
